move market api call to main page
